+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/completion/CompletionProviderUtils.java
@@ -30,6 +30,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 final class CompletionProviderUtils {
 
@@ -111,11 +112,14 @@ final class CompletionProviderUtils {
                                                     @NotNull CompletionResultSet resultSet) {
         String lookup = lookupElementMarker.getLookup();
         if (lookup != null) {
-            List<String> lookupStrings = List.of(lookup, WordUtils.capitalize(lookup), lookup.toLowerCase());
+            String[] lookupWords = lookupElementMarker.getLookupWords();
+            List<String> lookupStrings = Stream.concat(Stream.of(lookup), Arrays.stream(lookupWords)).toList();
             LookupElementBuilder builder = LookupElementBuilder.create(lookup)
                                                                .withLookupStrings(lookupStrings)
                                                                .withIcon(icon)
-                                                               .withPsiElement(lookupElementMarker.reference());
+                                                               .withPsiElement(lookupElementMarker.reference())
+                                                               .withCaseSensitivity(lookupElementMarker.isCaseSensitive())
+                                                               .withPresentableText(lookupElementMarker.getPresentableText());
             if (bold) {
                 builder = builder.bold();
             }